Refactor Main.dart
